+++ packages/sitecore-jss-dev-tools/src/index.ts
@@ -30,6 +30,7 @@ export {
   DisconnectedServerOptions,
 } from './disconnected-server/create-default-disconnected-server';
 export { ScJssConfig, JssConfiguration, resolveScJssConfig } from './resolve-scjssconfig';
+export { Metadata } from '@sitecore-jss/sitecore-jss/utils';
 
 export * from './templating';
 export * from './manifest';